requirement.txt
django==2.2.12
djangorestframework==3.12.2
django-rest-swagger==2.2.0
graphene-django==2.15.0
django-debug-toolbar==3.2
django-filter==2.4.0
psycopg2-binary==2.8.6
Pillow==8.1.0